Express explorer uis

+++ scripts/temp-fixes-after-install.sh
@@ -37,21 +37,35 @@ echo "[${SCRIPT_NAME}] change ${CI_ZOWE_DS_MEMBER} RACF user ..."
 echo
 
 ################################################################################
+# Error during zowe-install:
+# Exporting certificate zOSMFCA from z/OSMF:
+# keytool error (likely untranslated): java.io.FileNotFoundException: /zaas1/zowe-install/extracted/zowe-0.9.5/install/../temp_2018-12-19/zosmf_cert_zOSMFCA.cer (EDC5111I Permission denied.)
+# FIXME: su doesn't work well here
+# echo
+# echo "[${SCRIPT_NAME}] import z/OSMF certificates which requires superuser permission ..."
+# (exec sh -c "cd ${CI_ZOWE_ROOT_DIR}/api-mediation && su && export PATH=\$ZOWE_JAVA_HOME/bin:\$PATH && scripts/apiml_cm.sh --action trust-zosmf --zosmf-keyring IZUKeyring.IZUDFLT --zosmf-userid IZUSVR")
+# echo
+
+################################################################################
 # explorer JES/MVS/USS has internal host name, convert to public domain
 echo
 echo "[${SCRIPT_NAME}] checking hostname in explorer-* ..."
 ZDNT_HOSTNAME=S0W1
-FILES_TO_UPDATE="explorer-JES explorer-USS explorer-MVS api_catalog"
+FILES_TO_UPDATE="explorer-JES explorer-USS explorer-MVS api_catalog jes_explorer mvs_explorer uss_explorer"
 for one in $FILES_TO_UPDATE; do
   ZDNT_FILE=$CI_ZOWE_ROOT_DIR/${one}/web/index.html
   echo "[${SCRIPT_NAME}]   - checking $ZDNT_FILE ..."
-  HAS_WRONG_HOSTNAME=$(grep $ZDNT_HOSTNAME $ZDNT_FILE)
-  if [ -n "$HAS_WRONG_HOSTNAME" ]; then
-    sed "s#//${ZDNT_HOSTNAME}:\([0-9]\+\)/#//${CI_HOSTNAME}:\1/#" $ZDNT_FILE > index.html.tmp
-    mv index.html.tmp $ZDNT_FILE
-    echo "[${SCRIPT_NAME}]     - updated."
+  if [ -f "$ZDNT_FILE" ]; then
+    HAS_WRONG_HOSTNAME=$(grep $ZDNT_HOSTNAME $ZDNT_FILE)
+    if [ -n "$HAS_WRONG_HOSTNAME" ]; then
+      sed "s#//${ZDNT_HOSTNAME}:\([0-9]\+\)/#//${CI_HOSTNAME}:\1/#" $ZDNT_FILE > index.html.tmp
+      mv index.html.tmp $ZDNT_FILE
+      echo "[${SCRIPT_NAME}]     - updated."
+    else
+      echo "[${SCRIPT_NAME}]     - no need to update."
+    fi
   else
-    echo "[${SCRIPT_NAME}]     - no need to update."
+    echo "[${SCRIPT_NAME}]     - doesn't exist."
   fi
 done
